Finish back-end Investment

+++ backend/banking-backend/src/main/java/com/lucas/banking/banking_backend/dto/FinancialAssetDTO.java
@@ -2,12 +2,14 @@ package com.lucas.banking.banking_backend.dto;
 
 import com.lucas.banking.banking_backend.entity.FinancialAssetType;
 
+import java.math.BigDecimal;
+
 public record FinancialAssetDTO(
     String name,
     String ticker,
     FinancialAssetType type,
-    Double currentPrice,
-    Double yieldPercentage
+    BigDecimal currentPrice,
+    BigDecimal yieldPercentage
 ) {
     
 }
